llm_council/requirements.txt
openai
anthropic
google-genai
pydantic
python-dotenv
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
